Reorganize deepbook_query and deepbook_sdk into client, router, and transactions
