+++ sdk/deepbook/src/index.ts
@@ -1,5 +1,6 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './deepbook_query';
-export * from './deepbook_sdk';
+export * from './client';
+export * from './transactions';
+export * from './router';
